+++ src/models/severity_level.rs
@@ -8,4 +8,6 @@ pub(crate) enum SeverityLevel {
     Information = 1,
     Warning = 2,
     Error = 3,
+    #[cfg(feature = "logs")]
+    Critical = 4,
 }